Add rankFiD template and use the template config dataclass everywhere (#247)
- This cl adds the rankFiD template and integrates it to rankFiD model coordinator classes. We had to opt in for a yet new method since rankFiD was not compatible with neither multiturn nor singleturn listwise templates.
- It also changes template validations to use a dataclass instead of a dict.
- It update vicuna and zephyr reranker classes and some other minor updates.
