+++ src/rank_llm/rerank/listwise/rank_fid.py
@@ -33,7 +33,7 @@ class RankFiDDistill(ListwiseRankLLM):
         model: str,
         context_size: int = 150,
         prompt_mode: PromptMode = PromptMode.LiT5,  # Placeholder for actual mode
-        prompt_template_path: Optional[str] = None,
+        prompt_template_path: str = "src/rank_llm/rerank/prompt_templates/rank_fid_template.yaml",
         num_few_shot_examples: int = 0,
         few_shot_file: Optional[str] = None,
         window_size: int = 20,
@@ -188,20 +188,13 @@ class RankFiDDistill(ListwiseRankLLM):
         """
         Create a prompt based on the result and given ranking range.
         """
-
-        # For now, we concat the prompt, because it seems LiT5 is also concatting the stuff
-        prompts = [
-            {
-                "text": self._gen_passage(
-                    result.query.text,
-                    i + 1 - rank_start,
-                    self.convert_doc_to_prompt_content(
-                        result.candidates[i].doc, self.max_tokens()
-                    ),
-                )
-            }
-            for i in range(rank_start, rank_end)
-        ]
+        # TODO (issue #237): Need to modify this to add fewshot examples
+        prompts = self._inference_handler.generate_prompt(
+            result=result,
+            rank_start=rank_start,
+            rank_end=rank_end,
+            max_tokens=self.max_tokens(),
+        )
 
         return prompts, sum(self.get_num_tokens(prompt["text"]) for prompt in prompts)
 
@@ -246,10 +239,6 @@ class RankFiDDistill(ListwiseRankLLM):
 
             return output_token_estimate
 
-    @staticmethod
-    def _gen_passage(query: str, index: int, passage: str) -> str:
-        return f"Search Query: {query} Passage: [{index}] {passage} Relevance Ranking: "
-
 
 class RankFiDScore(ListwiseRankLLM):
     def _post_init(self):
@@ -276,7 +265,7 @@ class RankFiDScore(ListwiseRankLLM):
         model: str,
         context_size: int = 150,
         prompt_mode: PromptMode = PromptMode.LiT5,  # Placeholder for actual mode
-        prompt_template_path: Optional[str] = None,
+        prompt_template_path: str = "src/rank_llm/rerank/prompt_templates/rank_fid_score_template.yaml",
         num_few_shot_examples: int = 0,
         few_shot_file: Optional[str] = None,
         window_size: int = 20,
@@ -454,26 +443,15 @@ class RankFiDScore(ListwiseRankLLM):
         """
         Create a prompt based on the result and given ranking range.
         """
-        query = result.query.text
-        results = []
+        # TODO (issue #237): Need to modify this to add fewshot examples
+        prompts = self._inference_handler.generate_prompt(
+            result=result,
+            rank_start=rank_start,
+            rank_end=rank_end,
+            max_tokens=self.max_tokens(),
+        )
 
-        sum_token = 0
-
-        for i in range(rank_start, rank_end):
-            results.append(
-                {
-                    "query": f"question: {query}",
-                    "text": self._gen_passage(
-                        query,
-                        self.convert_doc_to_prompt_content(
-                            result.candidates[i].doc, self.max_tokens()
-                        ),
-                    ),
-                }
-            )
-            sum_token += len(self._tokenizer.encode(results[-1]["text"]))
-
-        return results, sum_token
+        return prompts, sum(self.get_num_tokens(prompt["text"]) for prompt in prompts)
 
     def get_num_tokens(self, prompt: str) -> int:
         return len(self._tokenizer.encode(prompt))
@@ -505,7 +483,3 @@ class RankFiDScore(ListwiseRankLLM):
                 self._output_token_estimate = output_token_estimate
 
             return output_token_estimate
-
-    @staticmethod
-    def _gen_passage(query: str, passage: str) -> str:
-        return f"question: {query} context: {passage}"
