+++ src/rank_llm/rerank/prompt_templates/rank_fid_score_template.yaml
@@ -1,0 +1,3 @@
+method: "rankfid"
+query: "question: {query}"
+text: "question: {query} context: {passage}"
