+++ src/rank_llm/rerank/prompt_templates/rank_fid_template.yaml
@@ -1,0 +1,2 @@
+method: "rankfid"
+text: "Search Query: {query} Passage: [{index}] {passage} Relevance Ranking: "
