+++ src/rank_llm/rerank/rankllm.py
@@ -209,6 +209,9 @@ class RankLLM(ABC):
         from rank_llm.rerank.listwise.multiturn_listwise_inference_handler import (
             MultiTurnListwiseInferenceHandler,
         )
+        from rank_llm.rerank.listwise.rankfid_inference_handler import (
+            RankFIDInferenceHandler,
+        )
         from rank_llm.rerank.listwise.singleturn_listwise_inference_handler import (
             SingleTurnListwiseInferenceHandler,
         )
@@ -224,6 +227,8 @@ class RankLLM(ABC):
                 return SingleTurnListwiseInferenceHandler(template)
             elif template["method"] == "multiturn_listwise":
                 return MultiTurnListwiseInferenceHandler(template)
+            elif template["method"] == "rankfid":
+                return RankFIDInferenceHandler(template)
             elif template["method"] == "pointwise":
                 return PointwiseInferenceHandler(template)
             elif template["method"] == "pairwise":
